+++ app/models/lab/lab_acknowledgement.rb
@@ -2,6 +2,5 @@
 
 module Lab
   class LabAcknowledgement < ::LimsAcknowledgementStatus
-    enum acknowledgment_type: %i[test_results_delivered_to_site_electronically test_results_delivered_to_site_manually]
   end
 end
